refactor: centralize utility helpers
